1.0.1 igoweb 990 total reads overflow (#8)
* [IGOWEB-990] Change number of reads from int to long

* [IGOWEB-990] Change number of reads from int to long for Sample

+++ src/main/java/org/mskcc/domain/sample/Sample.java
@@ -17,7 +17,7 @@ public class Sample {
     private Sample pairing;
     private String cmoSampleId = Constants.UNDEFINED;
     private final Predicate<Sample> validSamplePredicate;
-    private int numberOfReads;
+    private long numberOfReads;
     private String alias = "";
     private Map<String, String> properties = new LinkedHashMap<>();
     private boolean pooledNormal;
@@ -41,11 +41,11 @@ public class Sample {
         this.validSamplePredicate = validSamplePredicate;
     }
 
-    public int getNumberOfReads() {
+    public long getNumberOfReads() {
         return numberOfReads;
     }
 
-    public void setNumberOfReads(int numberOfReads) {
+    public void setNumberOfReads(long numberOfReads) {
         this.numberOfReads = numberOfReads;
     }
